Extract API clients to individual files.
Also extract all error classes into a dedicated object.
