+++ ota-plus-web/app/com/advancedtelematic/api/ApiClient.scala
@@ -2,36 +2,21 @@ package com.advancedtelematic.api
 
 import java.util.UUID
 
-import akka.Done
-import akka.stream.Materializer
 import akka.util.ByteString
 import brave.play.{TraceData, TraceWSClient, ZipkinTraceServiceLike}
 import com.advancedtelematic.api.ApiRequest.UserOptions
-import com.advancedtelematic.auth.AccessToken
-import com.advancedtelematic.controllers.{FeatureName, PathBinders, UserId}
+import com.advancedtelematic.controllers.FeatureName
 import com.advancedtelematic.libats.data.DataType.Namespace
-import com.advancedtelematic.libtuf.data.TufCodecs.tufKeyPairDecoder
-import com.advancedtelematic.libtuf.data.TufDataType.TufKeyPair
-import play.api.Configuration
-import play.api.http.Status
 import play.api.libs.json._
 import play.api.libs.ws.{BodyWritable, InMemoryBody, WSAuthScheme, WSClient, WSRequest, WSResponse}
 import play.api.mvc._
-import PathBinders.segment
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Try
-import scala.util.control.NoStackTrace
+
+import scala.concurrent.Future
 
 object ApiVersion extends Enumeration {
   type ApiVersion = Value
   val v1, v2 = Value
 }
-
-case class RemoteApiIOError(cause: Throwable) extends Exception(cause) with NoStackTrace
-
-case class RemoteApiError(result: Result, msg: String = "") extends Exception(msg) with NoStackTrace
-
-case class RemoteApiParseError(msg: String) extends Exception(msg) with NoStackTrace
 
 case class UserPass(user: String, pass: String)
 
@@ -123,248 +108,4 @@ trait ApiRequest { self =>
   def execJson[T](apiExec: ApiClientExec)(implicit ev: Reads[T]): Future[T] = {
     apiExec.runApiJson[T](build)
   }
-}
-
-/**
-  * Controllers extending [[ApiClientSupport]] access [[AuthPlusApi]] endpoints using a singleton.
-  */
-class AuthPlusApi(val conf: Configuration, val apiExec: ApiClientExec)
-                 (implicit tracer: ZipkinTraceServiceLike) extends OtaPlusConfig {
-  /**
-    * The response is json for `com.advancedtelematic.authplus.client.ClientInformationResponse`
-    * for a ClientID:
-    * <ul>
-    *   <li>the ClientID was generated by Auth+ upon registering a DeviceID</li>
-    *   <li>the web-app persisted the association DeviceID -> ClientID</li>
-    * </ul>
-    */
-  def fetchClientInfo(clientID: UUID, token: AccessToken)(implicit traceData: TraceData): Future[JsValue] = {
-    ApiRequest.traced("auth-plus", authPlusApiUri.uri + s"/clients/${clientID.toString}")
-      .withToken(token.value)
-      .transform(_.withMethod("GET"))
-      .execJsonValue(apiExec)
-  }
-
-  def fetchSecret(clientID: UUID, token: AccessToken)
-                 (implicit ec: ExecutionContext, traceData: TraceData): Future[String] = {
-    fetchClientInfo(clientID, token).flatMap { parsed =>
-      val t2: Try[String] = Try((parsed \ "client_secret").validate[String].get)
-      Future.fromTry(t2)
-    }
-  }
-
-}
-
-class UserProfileApi(val conf: Configuration, val apiExec: ApiClientExec)(implicit tracer: ZipkinTraceServiceLike)
-  extends OtaPlusConfig {
-
-  import play.api.libs.functional.syntax._
-
-  private def userProfileRequest(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("user-profile", userProfileApiUri.uri + "/api/v1/" + path)
-
-  implicit val featureNameR: Reads[FeatureName] = Reads.StringReads.map(FeatureName)
-  implicit val featureNameW: Writes[FeatureName] = Writes.StringWrites.contramap(_.get)
-  implicit val namespaceR: Reads[Namespace] = Reads.StringReads.map(Namespace(_))
-
-  implicit val featureR: Reads[Feature] = {(
-    (__ \ "feature").read[FeatureName] and
-    (__ \ "client_id").readNullable[UUID] and
-    (__ \ "enabled").read[Boolean]
-  )(Feature.apply _)}
-
-  implicit val userOrganizationR: Reads[UserOrganization] = {
-    (
-      (__ \ "namespace").read[Namespace] and
-      (__ \ "name").read[String] and
-      (__ \ "isDefault").read[Boolean]
-    )(UserOrganization.apply _)
-  }
-
-  def getUser(userId: UserId)(implicit traceData: TraceData): Future[JsValue] =
-    userProfileRequest("users/" + segment(userId.id)).execJsonValue(apiExec)
-
-  def createUser(userId: UserId, name: String, email: String)(implicit traceData: TraceData): Future[JsValue] =
-    userProfileRequest(s"users/${userId.id}")
-      .transform(_.withMethod("POST"))
-      .transform(_.withBody(Json.obj("name" -> name, "email"-> email)))
-      .execJsonValue(apiExec)
-
-  def getFeature(namespace: Namespace, feature: FeatureName)(implicit traceData: TraceData): Future[Feature] =
-    userProfileRequest("organizations/" + segment(namespace.get) + "/features/" + feature.get)
-      .execJson[Feature](apiExec)
-
-  def getFeatures(namespace: Namespace)(implicit traceData: TraceData): Future[Seq[FeatureName]] =
-    userProfileRequest("organizations/" + segment(namespace.get) + "/features").execJson[Seq[FeatureName]](apiExec)
-
-  def activateFeature(namespace: Namespace, feature: FeatureName, clientId: UUID)
-                     (implicit traceData: TraceData): Future[Result] = {
-    val requestBody = Json.obj("feature" -> feature.get, "client_id" -> clientId)
-
-    userProfileRequest(s"organizations/${segment(namespace.get)}/features")
-      .transform(_.withMethod("POST").withBody(requestBody))
-      .execResult(apiExec)
-  }
-
-  def updateDisplayName(userId: UserId, newName: String)
-                       (implicit executionContext: ExecutionContext, traceData: TraceData): Future[Done] = {
-    val request = userProfileRequest(s"users/${segment(userId.id)}/displayname").transform(
-      _.withMethod("PUT").withBody(JsString(newName))
-    ).build
-    apiExec.runSafe(request).map(_ => Done)
-  }
-
-  def updateBillingInfo[T](userId: UserId, query: Map[String,Seq[String]], body: JsValue)
-                          (implicit traceData: TraceData): Future[Result] =
-    userProfileRequest(s"users/${segment(userId.id)}/billing_info")
-      .transform(
-        _.withMethod("PUT")
-          .withQueryStringParameters(query.mapValues(_.head).toSeq :_*)
-          .withBody(body))
-      .execResult(apiExec)
-
-  def userProfileRequest(userId: UserId, method: String, path: String)(implicit traceData: TraceData): Future[Result] =
-    userProfileRequest(s"users/${segment(userId.id)}/$path")
-      .transform(_.withMethod(method))
-      .execResult(apiExec)
-
-  def userOrganizations(userId: UserId)(implicit traceData: TraceData): Future[Set[UserOrganization]] =
-    userProfileRequest(s"users/${segment(userId.id)}/organizations")
-      .transform(_.withMethod("GET"))
-      .execJson[Set[UserOrganization]](apiExec)
-
-  def namespaceIsAllowed(userId: UserId, namespace: Namespace)
-                        (implicit ec: ExecutionContext, traceData: TraceData): Future[Boolean] =
-    userOrganizations(userId).map(_.map(_.namespace)).map(_.contains(namespace))
-
-  def organizationRequest(namespace: Namespace, method: String, path: String, body: Option[JsValue])
-                         (implicit traceData: TraceData): Future[Result] =
-    userProfileRequest(s"organizations/${segment(namespace.get)}/$path")
-      .transform(_.withMethod(method))
-      .transform(r => body.map(json => r.withBody(json)).getOrElse(r))
-      .execResult(apiExec)
-
-  def getCredentialsBundle(namespace: Namespace, keyUuid: UUID)(implicit traceData: TraceData): Future[Result] =
-    userProfileRequest(s"${segment(namespace.get)}/credentials/$keyUuid")
-      .transform(_.withMethod("GET"))
-      .execResult(apiExec)
-}
-
-class DeviceRegistryApi(val conf: Configuration, val apiExec: ApiClientExec)
-                       (implicit tracer: ZipkinTraceServiceLike) extends OtaPlusConfig with CirceJsonBodyWritables {
-
-  private def request(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("device-registry", devicesApiUri.uri + "/api/v1/" + path)
-
-  def recentDevices(namespace: Namespace, limit: Int)(implicit traceData: TraceData): Future[JsValue] =
-    request("devices")
-      .transform(_.addQueryStringParameters("sortBy" -> "createdAt", "limit" -> limit.toString))
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-
-  def recentDeviceGroups(namespace: Namespace, limit: Int)(implicit traceData: TraceData): Future[JsValue] =
-    request("device_groups")
-      .transform(_.addQueryStringParameters("sortBy" -> "createdAt", "limit" -> limit.toString))
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-
-  def countDevicesInGroup(namespace: Namespace, groupId: String)(implicit traceData: TraceData): Future[JsValue] =
-    request(s"device_groups/$groupId/count")
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-}
-
-class CampaignerApi(val conf: Configuration, val apiExec: ApiClientExec)
-                   (implicit tracer: ZipkinTraceServiceLike) extends OtaPlusConfig with CirceJsonBodyWritables {
-
-  private def request(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("campaigner", campaignerApiUri.uri + "/api/v2/" + path)
-
-  def recentCampaigns(namespace: Namespace, limit: Int)(implicit traceData: TraceData): Future[JsValue] =
-    request("campaigns")
-      .transform(_.addQueryStringParameters("sortBy" -> "createdAt", "limit" -> limit.toString))
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-
-  def countDevicesInCampaign(namespace: Namespace, campaignId: String)
-                            (implicit traceData: TraceData, ec: ExecutionContext): Future[JsValue] =
-    request(s"campaigns/$campaignId/stats")
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-      .map(j => (j \ "processed").as[JsValue])
-
-  def recentUpdates(namespace: Namespace, limit: Int)(implicit traceData: TraceData): Future[JsValue] =
-    request("updates")
-      .transform(_.addQueryStringParameters("sortBy" -> "createdAt", "limit" -> limit.toString))
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-}
-
-class DirectorApi(val conf: Configuration, val apiExec: ApiClientExec)
-                 (implicit tracer: ZipkinTraceServiceLike) extends OtaPlusConfig with CirceJsonBodyWritables {
-
-  private def request(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("director", directorApiUri.uri + "/api/v1/" + path)
-
-  def fetchEcuTypes(namespace: Namespace, updateId: String)
-                   (implicit traceData: TraceData, ec: ExecutionContext): Future[JsValue] =
-    request(s"multi_target_updates/$updateId")
-      .withNamespace(Some(namespace))
-      .execJson[JsObject](apiExec)
-      .map(_.keys.toSeq.sorted.map(JsString))
-      .map(JsArray(_))
-}
-
-class RepoServerApi(val conf: Configuration, val apiExec: ApiClientExec)(implicit tracer: ZipkinTraceServiceLike)
-  extends OtaPlusConfig with CirceJsonBodyWritables {
-
-  private def request(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("reposerver", repoApiUri.uri + "/api/v1/" + path)
-
-  def rootJsonResult(namespace: Namespace)(implicit traceData: TraceData): Future[Result] =
-    request("user_repo/root.json")
-      .withNamespace(Some(namespace))
-      .execResult(apiExec)
-
-  def fetchTargets(namespace: Namespace)(implicit traceData: TraceData): Future[JsValue] =
-    request("user_repo/targets.json")
-      .withNamespace(Some(namespace))
-      .execJsonValue(apiExec)
-}
-
-class KeyServerApi(val conf: Configuration, val apiExec: ApiClientExec)
-                  (implicit tracer: ZipkinTraceServiceLike) extends OtaPlusConfig {
-  private def request(path: String)(implicit traceData: TraceData) =
-    ApiRequest.traced("keyserver", keyServerApiUri.uri + "/api/v1/" + path)
-
-  def targetKeys(repoId: String)
-                (implicit ec: ExecutionContext, mat: Materializer, traceData: TraceData): Future[Seq[TufKeyPair]] = {
-    // using circe since there is a decoder for it in the lib
-    import io.circe.{parser => CirceParser}
-
-    for {
-      result <- request(s"root/$repoId/keys/targets/pairs").execResult(apiExec)
-      byteString <- result.body.consumeData
-    } yield {
-      if (result.header.status == Status.NOT_FOUND) {
-        Seq.empty
-      } else {
-        val parsed = CirceParser.parse(byteString.utf8String) match {
-          case Left(t) =>
-            throw RemoteApiParseError(s"error parsing target keys (${byteString.utf8String}): ${t.message}")
-          case Right(json) =>
-            json
-        }
-
-        val vector = parsed.asArray.getOrElse(throw new Exception("Vector expected"))
-        vector.map { json =>
-          tufKeyPairDecoder.decodeJson(json) match {
-            case Left(t) => throw t
-            case Right(keyPair) => keyPair
-          }
-        }
-      }
-    }
-  }
-
 }
