--- ota-plus-web/app/com/advancedtelematic/api/MalformedResponse.scala
@@ -1,5 +1,0 @@
-package com.advancedtelematic.api
-
-import play.api.libs.ws.WSResponse
-
-final case class MalformedResponse(msg: String, response: WSResponse) extends Throwable(msg)
