--- ota-plus-web/app/com/advancedtelematic/api/UnexpectedResponse.scala
@@ -1,8 +1,0 @@
-package com.advancedtelematic.api
-
-import play.api.libs.ws.WSResponse
-
-final case class UnexpectedResponse(response: WSResponse) extends Throwable {
-  override def getMessage: String =
-    s"Unexpected response with status '${response.statusText}', body: ${response.body}"
-}
